requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asttokens==2.4.1
    # via icecream
blinker==1.7.0
    # via flask
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   pip-tools
colorama==0.4.6
    # via icecream
executing==2.0.1
    # via icecream
flask==3.0.2
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
icecream==2.1.3
    # via -r requirements.in
idna==3.6
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==23.2
    # via
    #   build
    #   gunicorn
pip-tools==7.3.0
    # via -r requirements.in
pygments==2.17.2
    # via icecream
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
ts3==1.0.11
    # via -r requirements.in
urllib3==2.2.0
    # via requests
werkzeug==3.0.1
    # via flask
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
